[Refactor] Move LargeIntValue from runtime to types (#53598)
