+++ be/src/column/fixed_length_column_base.cpp
@@ -19,8 +19,8 @@
 #include "column/vectorized_fwd.h"
 #include "exec/sorting/sort_helper.h"
 #include "gutil/casts.h"
-#include "runtime/large_int_value.h"
 #include "storage/decimal12.h"
+#include "types/large_int_value.h"
 #include "util/hash_util.hpp"
 #include "util/mysql_row_buffer.h"
 #include "util/value_generator.h"
