+++ be/src/exprs/cast_expr.h
@@ -24,8 +24,8 @@
 #include "column/vectorized_fwd.h"
 #include "exprs/column_ref.h"
 #include "exprs/expr.h"
-#include "runtime/large_int_value.h"
 #include "runtime/types.h"
+#include "types/large_int_value.h"
 
 namespace starrocks {
