+++ be/src/exprs/string_functions.cpp
@@ -43,9 +43,9 @@
 #include "gutil/strings/fastmem.h"
 #include "gutil/strings/strip.h"
 #include "gutil/strings/substitute.h"
-#include "runtime/large_int_value.h"
 #include "runtime/runtime_state.h"
 #include "storage/olap_define.h"
+#include "types/large_int_value.h"
 #include "util/raw_container.h"
 #include "util/sm3.h"
 #include "util/utf8.h"
