+++ be/src/storage/types.cpp
@@ -37,7 +37,6 @@
 #include "gutil/strings/numbers.h"
 #include "runtime/datetime_value.h"
 #include "runtime/decimalv2_value.h"
-#include "runtime/large_int_value.h"
 #include "runtime/mem_pool.h"
 #include "runtime/time_types.h"
 #include "storage/collection.h"
@@ -50,6 +49,7 @@
 #include "storage/type_traits.h"
 #include "types/array_type_info.h"
 #include "types/date_value.hpp"
+#include "types/large_int_value.h"
 #include "types/map_type_info.h"
 #include "types/struct_type_info.h"
 #include "util/hash_util.hpp"
