+++ be/src/types/CMakeLists.txt
@@ -20,6 +20,7 @@ add_library(Types STATIC
     date_value.cpp
     hll.cpp
     hll_sketch.cpp
+    large_int_value.cpp
     logical_type.cpp
     map_type_info.cpp
     struct_type_info.cpp
