+++ be/src/util/arrow/starrocks_column_to_arrow.cpp
@@ -21,8 +21,8 @@
 #include "common/statusor.h"
 #include "exec/arrow_type_traits.h"
 #include "exprs/expr.h"
-#include "runtime/large_int_value.h"
 #include "runtime/types.h"
+#include "types/large_int_value.h"
 #include "util/raw_container.h"
 
 namespace starrocks {
