+++ be/test/runtime/decimalv3_test.cpp
@@ -20,7 +20,7 @@
 #include <tuple>
 #include <vector>
 
-#include "runtime/large_int_value.h"
+#include "types/large_int_value.h"
 #include "util/logging.h"
 namespace starrocks {
