+++ be/test/runtime/large_int_value_test.cpp
@@ -15,17 +15,13 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "runtime/large_int_value.h"
+#include "types/large_int_value.h"
 
 #include <gtest/gtest.h>
 
-#include <boost/lexical_cast.hpp>
-#include <iostream>
-#include <sstream>
 #include <string>
 
 #include "common/configbase.h"
-#include "common/logging.h"
 #include "types/constexpr.h"
 
 namespace starrocks {
