+++ be/test/storage/rowset/frame_of_reference_page_test.cpp
@@ -41,12 +41,12 @@
 #include "column/column_helper.h"
 #include "column/column_viewer.h"
 #include "gutil/int128.h"
-#include "runtime/large_int_value.h"
 #include "runtime/mem_pool.h"
 #include "storage/chunk_helper.h"
 #include "storage/rowset/options.h"
 #include "storage/rowset/page_builder.h"
 #include "storage/rowset/page_decoder.h"
+#include "types/large_int_value.h"
 #include "util/logging.h"
 
 using starrocks::PageBuilderOptions;
